add pics, map thymeleaf, complete home page

+++ ecommerce/src/main/java/cybersoft/javabackend/java11/ecommerce/product/dto/UpdateProductDto.java
@@ -16,7 +16,7 @@ public class UpdateProductDto {
 
 	private String description;
 	
-	private long price;
+	private int price;
 
 	private int unitsInStock;
